@@ -4,10 +4,10 @@ description: >-
   dstore.
 homepage: https://github.com/chandu0101/dstore
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
   meta:
   
 
 dev_dependencies:
-    
+  pedantic:
