@@ -4,7 +4,7 @@ description: >-
   dstore.
 homepage: https://github.com/chandu0101/dstore
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
   analyzer:  # Pin the analyzer version to work around generatiuon failure
   build: 
@@ -12,8 +12,10 @@ dependencies:
   meta: 
   source_gen: 
   ansicolor: ^1.1.1
+  gql:
   dstore:
     path: ../dstore
 
 dev_dependencies:
+  pedantic: 
     
