@@ -4,7 +4,7 @@ description: >-
   dstore.
 homepage: https://github.com/chandu0101/dstore
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
   dstore:
     path: ../dstore
@@ -12,6 +12,7 @@ dependencies:
 
 dev_dependencies:
   build_runner:
+  pedantic:
   dstore_generator:
     path: ../dstore_generator
   json_serializable: 
