@@ -4,7 +4,7 @@ description: >-
   dstore.
 homepage: https://github.com/chandu0101/dstore
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
   flutter: 
     sdk: flutter
@@ -15,4 +15,5 @@ dependencies:
   
 
 dev_dependencies:
+  pedantic: 
     
