@@ -4,12 +4,13 @@ description: >-
   dstore.
 homepage: https://github.com/chandu0101/dstore
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 dependencies:
   dstore:
     path: ../dstore
   dio: 
 dev_dependencies:
+  pedantic:
   
       
 
