@@ -1,7 +1,5 @@
 import 'dart:async';
 
-import 'package:meta/meta.dart';
-
 import 'package:dstore/dstore.dart';
 import 'package:dstore/src/action.dart';
 
@@ -21,9 +19,9 @@ class AsyncActionField {
 }
 
 class StreamField<D> {
-  final StreamSubscription internalSubscription;
-  final D data;
-  final dynamic error;
+  final StreamSubscription? internalSubscription;
+  final D? data;
+  final dynamic? error;
   final bool loading;
   final bool completed;
 
@@ -35,11 +33,11 @@ class StreamField<D> {
       this.error});
 
   StreamField<D> copyWith({
-    StreamSubscription internalSubscription,
-    D data,
-    dynamic error,
-    bool loading,
-    bool completed,
+    StreamSubscription? internalSubscription,
+    D? data,
+    dynamic? error,
+    bool? loading,
+    bool? completed,
   }) {
     return StreamField<D>(
       internalSubscription: internalSubscription ?? this.internalSubscription,
@@ -61,14 +59,14 @@ class PStateMeta<S extends PStateModel> {
   final ReducerFn<S> reducer;
   final AReducerFn<S> aReducer;
   final S Function() ds;
-  final Map<String, dynamic> Function(S) serialize;
-  final S Function(Map<String, dynamic>) deserialize;
+  final Map<String, dynamic> Function(S)? serialize;
+  final S Function(Map<String, dynamic>)? deserialize;
 
   PStateMeta(
-      {@required this.aReducer,
-      @required this.group,
+      {required this.aReducer,
+      required this.group,
       this.serialize,
       this.deserialize,
-      @required this.reducer,
-      @required this.ds});
+      required this.reducer,
+      required this.ds});
 }
