@@ -1,9 +1,7 @@
-import 'package:meta/meta.dart';
-
 abstract class PersitantStorage {
   Future<void> init();
   bool get isInitialized;
-  Future<void> set({@required String key, @required Map<String, dynamic> json});
+  Future<void> set({required String key, required Map<String, dynamic> json});
   Future<Map<String, dynamic>> get(String key);
 }
 
@@ -14,5 +12,5 @@ class StorageOptions {
   final StorageWriteMode writeMode;
 
   StorageOptions(
-      {@required this.storage, this.writeMode = StorageWriteMode.DISKLAST});
+      {required this.storage, this.writeMode = StorageWriteMode.DISKLAST});
 }
