@@ -1,10 +1,8 @@
-import 'dart:async';
-
 import 'package:dstore/dstore.dart';
-import 'package:meta/meta.dart';
 
 import 'package:dstore/src/action.dart';
 import 'package:dstore/src/selector.dart';
+import 'package:meta/meta.dart';
 
 typedef Dispatch = dynamic Function(Action action);
 
@@ -13,31 +11,31 @@ typedef Middleware<State extends AppStateI> = dynamic Function(
 
 typedef Callback = dynamic Function();
 
-typedef SelectorUnSubscribeFn = dynamic Function(UnSubscribeOptions options);
+typedef SelectorUnSubscribeFn = dynamic Function(UnSubscribeOptions? options);
 
 class Store<S extends AppStateI> {
   final Map<String, PStateMeta<PStateModel>> meta;
   final Map<String, List<_SelectorListener>> selectorListeners = {};
-  List<Dispatch> _dispatchers;
-  Map<String, String> _reducerGroupToStateKeyMap = {};
-  S _state;
+  late final List<Dispatch> _dispatchers;
+  final Map<String, String> _reducerGroupToStateKeyMap = {};
+  late S _state;
 
   Store(
-      {@required this.meta,
-      @required S Function() stateCreator,
+      {required this.meta,
+      required S Function() stateCreator,
       List<Middleware<S>> middlewares = const [],
-      S initialState}) {
+      S? initialState}) {
     middlewares.add(asyncMiddleware);
     _dispatchers = _createDispatchers(middlewares);
     _prepareNormalStore(stateCreator);
   }
 
-  get state => _state;
+  S get state => _state;
 
   void _prepareNormalStore(S Function() stateCreator) {
     final AppStateI s = stateCreator();
-    final Map<String, dynamic> map = {};
-    this.meta.forEach((key, rg) {
+    final map = <String, dynamic>{};
+    meta.forEach((key, rg) {
       _reducerGroupToStateKeyMap[rg.group] = key;
       map[key] = rg.ds;
     });
@@ -54,19 +52,19 @@ class Store<S extends AppStateI> {
   }
 
   dynamic _defaultDispatch(Action action) {
-    final sk = this._reducerGroupToStateKeyMap[action.group];
-    final psm = this.meta[sk];
+    final sk = _reducerGroupToStateKeyMap[action.group]!;
+    final psm = meta[sk]!;
     final gsMap = _state.toMap();
-    final currentS = gsMap[sk];
+    final currentS = gsMap[sk]!;
     var newS = currentS;
     if (action.isProcessed) {
       // processed by middlewares
-      if (action.internal.type == ActionInternalType.DATA) {
+      if (action.internal?.type == ActionInternalType.DATA) {
         final csMap = currentS.toMap();
-        csMap[action.name] = action.internal.data;
+        csMap[action.name] = action.internal!.data;
         newS = currentS.copyWithMap(csMap);
-      } else if (action.internal.type == ActionInternalType.STATE) {
-        newS = action.internal.data;
+      } else if (action.internal!.type == ActionInternalType.STATE) {
+        newS = action.internal!.data;
       }
     } else {
       if (action.isAsync) {
@@ -83,16 +81,16 @@ class Store<S extends AppStateI> {
   }
 
   void _notifyListeners(
-      {@required String stateKey,
-      @required PStateModel previousState,
-      @required PStateModel currentState}) {
-    final ls = this.selectorListeners[stateKey];
+      {required String stateKey,
+      required PStateModel previousState,
+      required PStateModel currentState}) {
+    final ls = selectorListeners[stateKey];
     final psMap = previousState.toMap();
     final csMap = currentState.toMap();
     if (ls != null) {
       ls.forEach((sl) {
         if (_isSelectorDependenciesChanged(
-            selector: sl.selector,
+            selector: sl.selector as Selector<S, dynamic>,
             prevState: psMap,
             currentState: csMap,
             stateKey: stateKey)) {
@@ -103,10 +101,10 @@ class Store<S extends AppStateI> {
   }
 
   bool _isSelectorDependenciesChanged(
-      {@required Selector<S, dynamic> selector,
-      @required Map<String, dynamic> prevState,
-      @required Map<String, dynamic> currentState,
-      @required String stateKey}) {
+      {required Selector<S, dynamic> selector,
+      required Map<String, dynamic> prevState,
+      required Map<String, dynamic> currentState,
+      required String stateKey}) {
     var result = false;
     selector.deps.forEach((key, value) {
       if (key == stateKey) {
@@ -126,16 +124,16 @@ class Store<S extends AppStateI> {
     final keysToReset = <String>[];
     final propsOfKeysToReset = <String, List<String>>{};
     selector.deps.forEach((sk, values) {
-      final slsa = this.selectorListeners[sk];
+      final slsa = selectorListeners[sk];
       if (slsa != null && slsa.length > 0) {
         final existingStateKeyProps = <String>{};
         slsa.forEach((sls) {
-          existingStateKeyProps.addAll(sls.selector.deps[sk]);
+          existingStateKeyProps.addAll(sls.selector.deps[sk]!);
         });
         propsOfKeysToReset[sk] = [];
         values.forEach((skp) {
           if (!existingStateKeyProps.contains(skp)) {
-            propsOfKeysToReset[sk].add(skp);
+            propsOfKeysToReset[sk]!.add(skp);
           }
         });
       } else {
@@ -144,13 +142,13 @@ class Store<S extends AppStateI> {
       }
       final sMap = _state.toMap();
       keysToReset.forEach((sk) {
-        sMap[sk] = this.meta[sk].ds();
+        sMap[sk] = meta[sk]!.ds();
       });
       propsOfKeysToReset.forEach((sk, props) {
         if (props.length > 0) {
-          final rm = sMap[sk];
+          final rm = sMap[sk]!;
           final rmMap = rm.toMap();
-          final rmDSMap = this.meta[sk].ds().toMap();
+          final rmDSMap = meta[sk]!.ds().toMap();
           props.forEach((prop) {
             rmMap[prop] = rmDSMap[prop];
           });
@@ -164,12 +162,12 @@ class Store<S extends AppStateI> {
   /* public methods  */
 
   String getStateKeyForReducerGroup(String key) {
-    return this._reducerGroupToStateKeyMap[key];
+    return _reducerGroupToStateKeyMap[key]!;
   }
 
   dynamic getFieldFromAction(Action action) {
-    final sk = this._reducerGroupToStateKeyMap[action.group];
-    final gsMap = this.state.toMap();
+    final sk = _reducerGroupToStateKeyMap[action.group];
+    final gsMap = state.toMap();
     final currentS = gsMap[sk] as PStateModel;
     return currentS.toMap()[action.name];
   }
@@ -179,24 +177,24 @@ class Store<S extends AppStateI> {
   }
 
   SelectorUnSubscribeFn subscribeSelector(
-      Selector<S, dynamic> selector, Callback listener) {
+      Selector<dynamic, dynamic> selector, Callback listener) {
     final keys = selector.deps.keys;
     keys.forEach((sk) {
-      final sls = this.selectorListeners[sk];
+      final sls = selectorListeners[sk];
       final v = _SelectorListener(selector: selector, listener: listener);
       if (sls != null) {
         sls.add(v);
       } else {
-        this.selectorListeners[sk] = [v];
+        selectorListeners[sk] = [v];
       }
     });
     var isSubscribed = true;
-    return ([UnSubscribeOptions options]) {
+    return ([UnSubscribeOptions? options]) {
       if (!isSubscribed) {
         return;
       }
       keys.forEach((sk) {
-        final sla = this.selectorListeners[sk];
+        final sla = selectorListeners[sk]!;
         final index = sla.indexWhere(
             (sl) => sl.selector == selector && sl.listener == listener);
         if (index >= 0) {
@@ -204,22 +202,24 @@ class Store<S extends AppStateI> {
         }
       });
       if (options != null && options.resetToDefault) {
-        _resetToDefaultStateIfNotUsedByOtherSelectors(selector);
+        _resetToDefaultStateIfNotUsedByOtherSelectors(
+            selector as Selector<S, dynamic>);
       } else {}
       isSubscribed = false;
     };
   }
 }
 
+// ignore: always_declare_return_types
 asyncMiddleware<S extends AppStateI>(
     Store<S> store, Dispatch next, Action action) async {
   if (action.isProcessed || !action.isAsync) {
     next(action);
   } else {
     final sk = store.getStateKeyForReducerGroup(action.group);
-    final psm = store.meta[sk];
+    final psm = store.meta[sk]!;
     final gsMap = store.state.toMap();
-    final currentS = gsMap[sk];
+    final currentS = gsMap[sk]!;
     store.dispatch(action.copyWith(
         internal: ActionInternal(
             processed: true,
@@ -243,12 +243,13 @@ asyncMiddleware<S extends AppStateI>(
   }
 }
 
+// ignore: always_declare_return_types
 streamMiddleware<S extends AppStateI>(
     Store<S> store, Dispatch next, Action action) async {
   if (action.isProcessed || action.stream == null) {
     next(action);
   } else {
-    final sub = action.stream.listen((event) {
+    final sub = action.stream?.listen((event) {
       final field = store.getFieldFromAction(action) as StreamField;
       store.dispatch(action.copyWith(
           internal: ActionInternal(
@@ -293,8 +294,8 @@ class _SelectorListener<S extends AppStateI> {
   final Selector selector;
   final Callback listener;
   _SelectorListener({
-    @required this.selector,
-    @required this.listener,
+    required this.selector,
+    required this.listener,
   });
 }
 
